refactor(playground): split chat input hook
