+++ packages/agent-playground/src/hooks/use-chat-input/index.ts
@@ -1,0 +1,8 @@
+export { useChatInput } from './use-chat-input';
+export type {
+  IChatInputHookReturn,
+  IChatInputOptions,
+  IChatInputState,
+  IChatMessage,
+  IInputValidationResult,
+} from './types';
